Add Policy for bounded distance placement
